requirements.txt
requests
GitPython
google-genai
dotenv
datetime